cql3: Implement ALTER INDEX
Implement a new CQL3 statement:

```cql
ALTER INDEX <KS>.<IDX> WITH (<property> = <value>)+
```

It allows for modifying the properties of the underlying materialized
view, and it follows the same restrictions as ALTER MATERIALIZED VIEW.
It does NOT allow for changing the properties of an index itself for
the following reasons:

* Currently, the spectrum of the index-specific options is not broad
  and even those that are used are used by vector indexes only.
* We have not decided whether we want to allow for altering the options
  of a vector index yet.
* Vector indexes are going to drop using the underlying materialized
  view, cf. VECTOR-10. Because of that, altering the options of a vector
  index deserves to be part of a separate set of changes.

We update the relevant parts of the developer documentation and provide
tests verifying that the changes work properly.

We will update the user documentation in a separate commit.