conf/scylla.yaml: Avoid explicitly setting value of rf_rack_valid_keyspaces
Although pretty much all of the instances of Scylla we use in our testing enable
the configuration option `rf_rack_valid_keyspaces` (scylladb/scylladb#23428),
we still have room for improvement.

Some of our tools reuse the same codepaths as Scylla, and they may rely on
its configuration. In particular, their behavior may depend on the value of
of `rf_rack_valid_keyspaces`.

Unfortunately, it's not as easy to enforce the option in the tools as in
instances of Scylla, so we need to cover them separately.

Production clusters might still use RF-rack-invalid keyspaces, and enabling
the option in the tools is not feasible as it could lead to undesired failures
that shouldn't occur.

One solution would be to parameterize the relevant tools and allow for toggling
the option. However, the option is not really related to what the tool does, so
it would be unnatural and confusing for the user.

The solution we're going to employ is to require that the user provides the
right value in the configuration file of Scylla. The tools will then make
decision based on it.

The first step is to avoid explicitly setting the option in the default
`scylla.yaml` file delivered along with Scylla to determine if the user
set the option to some value or not.

In the following commit, we're going to use that information to correctly
decide whether to enforce RF-rack-valid keyspaces in our tooling, or not.

The change has no direct impact on the user. The default value of the option
is `false`, and so we should observe no difference in the behavior of clusters.

The only indirect change is that the user may have to explicitly set the
option in the configuration file before using the tool.

As a bonus, we elaborate in `scylla.yaml` on what the option controls.