test/cluster/random_failures: Adjust to work with RF-rack-validity
We adjust the test to work with the configuration option
`rf_rack_valid_keyspaces` enabled. For that, we ensure that there is
always at least one node in each of the three racks. This way, all
keyspaces we create and manipulate will remain RF-rack-valid since they
all use RF=3.

------------------------------------------------------------------------

To achieve that, we only need to adjust the following events:

1. `init_tablet_transfer`
   The event creates a new keyspace and table and manually migrates
   a tablet belonging to it. As long as we make sure the migration occurs
   within the same rack, there will be no problem.

   Since RF == #racks, each rack will have exactly one tablet replica,
   so we can migrate the tablet to an arbitrary node in the same rack.

   Note that such a node must exist. If it didn't, the test wouldn't
   have worked before because we cannot migrate a tablet replica to
   another replica of that tablet. In other words, we have a guarantee
   that there are at least 4 nodes in the cluster when we try to migrate
   a tablet replica.

   That said, we check that. If there's no viable node to migrate the
   tablet replica to, we log that information and do nothing. Since the
   test is randomized and tries to test various problematic scenarios,
   that should be an acceptable solution.

2. `add_new_node`
   As long as we add a node to an existing rack, there's no way to
   violate the invariant imposed by the configuration option, so we pick
   a random rack out of the existing three and create a node in it.

3. `decommission_node`
   We need to ensure that the node we'll be trying to decommission is
   not the only one in its rack.

   Thankfully, such a rack must exist. If it could not, the test would
   already be broken: it's impossible to reduce the size of the cluster
   below the value of the replication factor with tablets; any attempt
   to decommission a node in a situation like that would result in
   a failure.

   What's more, decommissioning a node is not possible if any node in
   the cluster is dead, so we can assume that `manager.running_servers`
   returns the whole cluster.

4. `remove_node`
   The same as `decommission_node`. Just note that the node we choose to
   remove is first stopped, so again: the whole cluster must be returned
   by `manager.running_servers`.

------------------------------------------------------------------------

There's one more important thing to note. The test may sometimes trigger
a sequence of events where a new node is started, but, due to an error
injection, its initialization is not completed. Among other things, the
node may NOT have a host ID recognized by the rest of the nodes in the
cluster, and operations like tablet migration will fail if they target
it.

Thankfully, there seems to be a way to avoid problems stemming from
that. When a new node is added to the cluster, it should appear at the
end of the list returned by `manager.running_servers`. This most likely
stems from how dictionaries work in Python:

"Keys and values are iterated over in insertion order."
-- https://docs.python.org/3/library/stdtypes.html#dict-views

and the fact that we keep track of running servers using a dictionary.

Furthermore, we rely on the assumption that the test currently works
correctly. Assume that there is only one node per rack that has
a host ID recognized by the rest of the cluster and that take part
in the operations mentioned above. If that's the case, then
decommissioning any of the nodes with a tablet will fail because
the tablet will not have any viable target to migrate to. It can't
migrate to another replica after all, so the test must be wrong.
But since we assumed it's not, then we must always have at least
one rack with two "fully capable" nodes.

Using those assumptions, an auxiliary function, `select_viable_rack`,
was designed to carefully choose a correct rack, from which we'll then
pick nodes to perform the topological operations. It's simple: we just
find the first rack in the list that has at least two nodes in it.
That should ensure that we perform an operation that doesn't lead to
any unforeseen disaster.

------------------------------------------------------------------------

Since the test effectively becomes more complex due to more care for
the topology of the cluster, we extend the log messages to make them
more helpful when debugging a failure.