service/qos: Do not crash Scylla if auth_integration absent
If the user connects to Scylla via the maintenance socket, it may happen
that `auth_integration` has not been registered in the service level
controller yet. One example is maintenance mode when that will never
happen; another when the connection occurs before Scylla is fully
initialized.

To avoid unnecessary crashes, we reduce the severity of error in case
`auth_integration` is missing. Instead of crashing a node, we return
the default scheduling group or the empty optional, depending on the
situation. If the user tries to describe service levels, we throw
an exception.

That's important because when connecting to Scylla, we set the right
service level for the connection, which effectively involves leveraging
`auth_integration`. Thanks to the changes, using the maintenance socket
should be possible again.

We add a short test verifying that we can communicate with Scylla and
execute queries even when it's not been fully initialized yet.

Fixes scylladb/scylladb#26816